Add files via upload
This is the Planetary gear used for the SeaSkate mechanism. It should let us keep the sun in place while the 5 planets around it rotate. 
All planets, the sun, and the inner ring have been designed to keep all parts in place by adding a V-Shape pattern to the outer parts of each of them, that way the planets don't get unscrewed out of the mechanism by rotating. The outer ring was designed to be 5cm in diameter to keep the structure strong at the same time as cutting weight and price while 3D printing. The whole Truck was designed on top of this piece.